Fix : Refactoring
