+++ usecase/todo_repository.go
@@ -1,0 +1,8 @@
+package usecase
+
+import "github.com/halkn/echo-sample/entity"
+
+// TodoRepository ...
+type TodoRepository interface {
+	FindAll() (entity.Todos, error)
+}
