+++ usecase/todo_usecase.go
@@ -4,5 +4,5 @@ import "github.com/halkn/echo-sample/entity"
 
 // TodoUsecase have usecase for todo.
 type TodoUsecase interface {
-	FindAll() (entity.Todos, error)
+	GetAllTodos() (entity.Todos, error)
 }
